+++ src/main/kotlin/com/vk/admstorm/services/WatchDebugLogService.kt
@@ -5,26 +5,25 @@ import com.intellij.openapi.Disposable
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
-import com.vk.admstorm.executors.YarnWatchCommandExecutor
+import com.vk.admstorm.env.Env
+import com.vk.admstorm.executors.WatchDebugLogCommandExecutor
 import com.vk.admstorm.ssh.SshConnectionService
 import com.vk.admstorm.utils.MyUtils.executeOnPooledThread
-import com.vk.admstorm.utils.MyUtils.invokeAfter
 import java.beans.PropertyChangeSupport
 
 @Service
-class YarnWatchService(private val myProject: Project) : Disposable {
+class WatchDebugLogService(private val project: Project) : Disposable {
     companion object {
-        const val PROPERTY_ID = "admstorm.yarn.watch.running.state"
-        fun getInstance(project: Project) = project.service<YarnWatchService>()
+        const val PROPERTY_ID = "admstorm.watch.debug.log.running.state"
+        fun getInstance(project: Project) = project.service<WatchDebugLogService>()
     }
 
     enum class State {
         RUNNING,
         STOPPED,
-        WITH_ERRORS
     }
 
-    private var executor: YarnWatchCommandExecutor? = null
+    private var executor: WatchDebugLogCommandExecutor? = null
     val changes = PropertyChangeSupport(this)
 
     fun isRunning() = state() != State.STOPPED
@@ -42,7 +41,7 @@ class YarnWatchService(private val myProject: Project) : Disposable {
             return
         }
 
-        executor = YarnWatchCommandExecutor(myProject, "FORCE_COLOR=true yarn watch")
+        executor = WatchDebugLogCommandExecutor(project, "${Env.data.vkCommand} debug")
 
         executeOnPooledThread {
             if (executor!!.run()) {
@@ -60,28 +59,9 @@ class YarnWatchService(private val myProject: Project) : Disposable {
 
     fun showConsole() {
         executor?.showToolWindow()
-        clearErrorsState()
     }
 
-    fun state() = State.valueOf(PropertiesComponent.getInstance(myProject).getValue(PROPERTY_ID, State.STOPPED.name))
-
-    fun setErrorsState() {
-        // Если окно не открыто, то бесконечно показываем
-        // анимацию ошибки, если же открыто, то только 5
-        // секунд.
-        if (executor!!.isToolWindowVisible()) {
-            invokeAfter(5000) {
-                clearErrorsState()
-            }
-        }
-        setState(State.WITH_ERRORS)
-    }
-
-    fun clearErrorsState() {
-        if (state() == State.WITH_ERRORS) {
-            setState(State.RUNNING)
-        }
-    }
+    fun state() = State.valueOf(PropertiesComponent.getInstance(project).getValue(PROPERTY_ID, State.STOPPED.name))
 
     fun setWorking(value: Boolean) {
         if (value) {
@@ -92,11 +72,11 @@ class YarnWatchService(private val myProject: Project) : Disposable {
     }
 
     private fun setState(value: State) {
-        PropertiesComponent.getInstance(myProject).setValue(PROPERTY_ID, value.name)
+        PropertiesComponent.getInstance(project).setValue(PROPERTY_ID, value.name)
         changes.firePropertyChange(PROPERTY_ID, "", value.name)
     }
 
-    private fun isConnected() = SshConnectionService.getInstance(myProject).isConnectedOrWarning()
+    private fun isConnected() = SshConnectionService.getInstance(project).isConnectedOrWarning()
 
     override fun dispose() {
         executor?.dispose()